backend/apps/gallery/serializers.py
from rest_framework import serializers
from .models import GalleryItem


class GalleryItemSerializer(serializers.ModelSerializer):
    program_slug = serializers.CharField(source='program.slug', read_only=True)
    article_slug = serializers.CharField(source='article.slug', read_only=True)

    class Meta:
        model = GalleryItem
        fields = '__all__'
