backend/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
cloudinary
django-cloudinary-storage
dj-database-url
psycopg2-binary
django-ckeditor
python-dotenv
gunicorn
whitenoise
Pillow
six
